app/core/settings.py
from pydantic_settings import BaseSettings, SettingsConfigDict
from pydantic import Field
from dotenv import load_dotenv

# Ensure .env is loaded at import time (works for app/tests/cli)
load_dotenv()


class Settings(BaseSettings):
    # Required in all environments (tests will set it explicitly)
    secret_key: str = Field(...)
    algorithm: str = Field(default="HS256")
    access_token_expire_minutes: int = Field(default=60)

    model_config = SettingsConfigDict(env_file=".env", extra="ignore")


def get_settings() -> Settings:
    return Settings()
